requirements.txt
azure-functions==1.21.3
requests==2.31.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
